requirements.txt
ytmusicapi
PyYAML